fix(ui): preserve target order for landed steers
